lab6/requirements.txt
gdown
wandb
tqdm
matplotlib
pillow
numpy

